separating msmws installation
