+++ requirements.d/mswms.txt
@@ -1,0 +1,45 @@
+# This file may be used to create an environment using:
+# $ conda create --name <env> --file <this file>
+# platform: linux-64
+python
+future
+defusedxml
+basemap>=1.3.3
+chameleon
+execnet
+fastkml>=0.11
+shapely>=2.0.0
+pygeoif<1.0.0
+isodate
+lxml
+netcdf4
+hdf4
+pillow
+requests>=2.31.0
+scipy
+skyfield>=1.12
+skyfield-data>=5
+tk
+owslib>=0.24
+unicodecsv
+fs
+cftime>=1.0.1
+matplotlib-base>=3.5.3
+itsdangerous
+flask>=2.3.2
+flask-httpauth
+werkzeug>=2.2.3,<3.0.0
+passlib
+multidict
+pint
+markdown
+xstatic
+xstatic-jquery
+xstatic-bootstrap
+gpxpy>=1.4.2
+metpy
+pycountry
+libtiff<4.5.0
+flask-wtf
+python-slugify
+flask-login
